Fix the CMake-port-specific part of ticket #40648.
This reverts the changes made to the CMake Portfile in r110069 so that
ports, which are configured by CMake, honor the configure.optflags set
in the Portfile. Thus, Release builds are compiled with optimization
by default again.

The MacPorts-specific issue of handling configure.optflags has to be
resolved in the CMake PortGroup.