Change the CMAKE_BUILD_TYPE to Debug instead of debugFull for the +debug variant in the CMake PortGroup.
"debugFull" is not a standard build type in CMake so there are no
default compiler flags set for it, which would turn on the inclusion
of debug symbols. Instead, CMake provides the build type "Debug" for
this purpose. Only KDE provides in its CMake build infrastructure
support for a build type called "DebugFull", which turns off
optimization, as KDE leaves it on for the "Debug" build type. Confer
to <http://techbase.kde.org/Development/CMake/Addons_for_KDE> for
further explanations.